requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7